fix: validate nil repository_name condition entries
